move (#3)
* new state, town

* moving

* refactor websocket

* communicate move

* no assums init

* handle negative position

+++ src/core/position.rs
@@ -1,4 +1,4 @@
 pub struct Position {
-    pub x: f64,
-    pub y: f64,
+    pub x: u32,
+    pub y: u32,
 }
